add const to methods, fix conversions

+++ server/base64.cpp
@@ -59,7 +59,7 @@ std::string base64_encode(const std::vector<uint8_t>& buf)
 
 std::vector<uint8_t> base64_decode(const std::string& encoded_string)
 {
-    int in_len = encoded_string.size();
+    int in_len = static_cast<int>(encoded_string.size());
     int i = 0;
     int j = 0;
     int in_ = 0;
@@ -72,7 +72,7 @@ std::vector<uint8_t> base64_decode(const std::string& encoded_string)
         if (i == 4)
         {
             for (i = 0; i < 4; i++)
-                char_array_4[i] = base64_chars.find(char_array_4[i]);
+                char_array_4[i] = static_cast<uint8_t>(base64_chars.find(char_array_4[i]));
 
             char_array_3[0] = (char_array_4[0] << 2) + ((char_array_4[1] & 0x30) >> 4);
             char_array_3[1] = ((char_array_4[1] & 0xf) << 4) + ((char_array_4[2] & 0x3c) >> 2);
@@ -90,7 +90,7 @@ std::vector<uint8_t> base64_decode(const std::string& encoded_string)
             char_array_4[j] = 0;
 
         for (j = 0; j < 4; j++)
-            char_array_4[j] = base64_chars.find(char_array_4[j]);
+            char_array_4[j] = static_cast<uint8_t>(base64_chars.find(char_array_4[j]));
 
         char_array_3[0] = (char_array_4[0] << 2) + ((char_array_4[1] & 0x30) >> 4);
         char_array_3[1] = ((char_array_4[1] & 0xf) << 4) + ((char_array_4[2] & 0x3c) >> 2);
